cleanup + test fixes
