--- brainrender/morphology/morphology_cache.py
@@ -1,6 +1,0 @@
-import os
-
-from vtkplotter import write, load
-
-
-conda
